Second version of hybrid model
Model does not work. The SD is just the same as in the previous version (Modelo_hibrido_ballenas4) but I tried to link it to the ABM model of whales that I did for the Santa Fe course (the one that has boats degrading the patches' quality (i.e. the habitat), though patches can "recover").